+++ atcg_lib/include/Renderer/Framebuffer.h
@@ -45,16 +45,26 @@ public:
     void attachColor();
 
     /**
+     * @brief Add a multi sampled color attachement
+     */
+    void attachColorMultiSample(uint32_t num_samples);
+
+    /**
      * @brief Attach a texture to the framebuffer
      *
      * @param texture The texture to attach
      */
-    void attachTexture(const atcg::ref_ptr<Texture2D>& texture);
+    void attachTexture(const atcg::ref_ptr<Texture>& texture);
 
     /**
      * @brief Add a depth attachement
      */
     void attachDepth();
+
+    /**
+     * @brief Add a multi sampled depth attachement
+     */
+    void attachDepthMultiSample(uint32_t num_samples);
 
     /**
      * @brief Attach a custom depth component
@@ -79,7 +89,7 @@ public:
      * @param slot The number of which attachement to use
      * @return The specified texture
      */
-    ATCG_INLINE atcg::ref_ptr<Texture2D> getColorAttachement(const uint32_t& slot = 0) const
+    ATCG_INLINE atcg::ref_ptr<Texture> getColorAttachement(const uint32_t& slot = 0) const
     {
         return _color_attachements[slot];
     }
@@ -135,7 +145,7 @@ private:
     static uint32_t s_current_fbo;
     uint32_t _ID;
     uint32_t _width, _height;
-    std::vector<atcg::ref_ptr<Texture2D>> _color_attachements;
+    std::vector<atcg::ref_ptr<Texture>> _color_attachements;
     atcg::ref_ptr<Texture> _depth_attachement;
 };
 }    // namespace atcg