+++ atcg_lib/platform/opengl/src/Renderer/Renderer.cpp
@@ -42,6 +42,8 @@ public:
     void initCameraFrustrum();
     atcg::ref_ptr<Graph> camera_frustrum;
 
+    void initFramebuffer(uint32_t num_frames, uint32_t width, uint32_t height);
+
     Material standard_material;
 
     atcg::ref_ptr<Texture2D> skybox_texture;
@@ -52,6 +54,9 @@ public:
     bool has_skybox = false;
 
     atcg::ref_ptr<Framebuffer> screen_fbo;
+    atcg::ref_ptr<Framebuffer> screen_fbo_msaa;
+    uint32_t msaa_num_samples = 8;
+    bool msaa_enabled         = true;
 
     atcg::ref_ptr<Graph> sphere_mesh;
     atcg::ref_ptr<Graph> cylinder_mesh;
@@ -198,15 +203,7 @@ RendererSystem::Impl::Impl(uint32_t width, uint32_t height, const atcg::ref_ptr<
     spec_lut.sampler.wrap_mode = TextureWrapMode::CLAMP_TO_EDGE;
     lut                        = atcg::Texture2D::create(img, spec_lut);
 
-    screen_fbo = atcg::make_ref<Framebuffer>(width, height);
-    screen_fbo->attachColor();
-    TextureSpecification spec_int;
-    spec_int.width  = width;
-    spec_int.height = height;
-    spec_int.format = TextureFormat::RINT;
-    screen_fbo->attachTexture(Texture2D::create(spec_int));
-    screen_fbo->attachDepth();
-    screen_fbo->complete();
+    initFramebuffer(msaa_num_samples, width, height);
 
     int total_units;
     glGetIntegerv(GL_MAX_COMBINED_TEXTURE_IMAGE_UNITS, &total_units);
@@ -320,6 +317,27 @@ void RendererSystem::Impl::initCameraFrustrum()
     edges.push_back({glm::vec2(4, 1), glm::vec3(1), 0.01f});
 
     camera_frustrum = atcg::Graph::createGraph(points, edges);
+}
+
+void RendererSystem::Impl::initFramebuffer(uint32_t num_samples, uint32_t width, uint32_t height)
+{
+    ATCG_ASSERT(context->isCurrent(), "Context of Renderer not current.");
+
+    screen_fbo = atcg::make_ref<Framebuffer>(width, height);
+    screen_fbo->attachColor();
+    TextureSpecification spec_int;
+    spec_int.width  = width;
+    spec_int.height = height;
+    spec_int.format = TextureFormat::RINT;
+    screen_fbo->attachTexture(Texture2D::create(spec_int));
+    screen_fbo->attachDepth();
+    screen_fbo->complete();
+
+    screen_fbo_msaa = atcg::make_ref<Framebuffer>(width, height);
+    screen_fbo_msaa->attachColorMultiSample(num_samples);
+    screen_fbo_msaa->attachTexture(Texture2DMultiSample::create(num_samples, spec_int));
+    screen_fbo_msaa->attachDepthMultiSample(num_samples);
+    screen_fbo_msaa->complete();
 }
 
 void RendererSystem::Impl::setMaterial(const Material& material, const atcg::ref_ptr<Shader>& shader)
@@ -909,6 +927,7 @@ void RendererSystem::init(uint32_t width,
     glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
     glEnable(GL_TEXTURE_CUBE_MAP_SEAMLESS);
     glPixelStorei(GL_UNPACK_ALIGNMENT, 1);
+    glEnable(GL_MULTISAMPLE);
     setViewport(0, 0, width, height);
 
     impl->shader_manager = shader_manager;
@@ -945,6 +964,7 @@ void RendererSystem::finishFrame()
     shader->setInt("screen_texture", 0);
 
     shader->use();
+    if(impl->msaa_enabled) impl->screen_fbo->blit(impl->screen_fbo_msaa);
     impl->screen_fbo->getColorAttachement()->use();
 
     const atcg::ref_ptr<IndexBuffer> ibo = impl->quad_vao->getIndexBuffer();
@@ -963,6 +983,7 @@ void RendererSystem::finishFrame()
         glBindTexture(GL_TEXTURE_2D_ARRAY, 0);
         glBindTexture(GL_TEXTURE_CUBE_MAP, 0);
         glBindTexture(GL_TEXTURE_CUBE_MAP_ARRAY, 0);
+        glBindTexture(GL_TEXTURE_2D_MULTISAMPLE, 0);
     }
     glActiveTexture(GL_TEXTURE0);
 }
@@ -991,6 +1012,22 @@ void RendererSystem::setLineSize(const float& size)
     ATCG_ASSERT(impl->context->isCurrent(), "Context of Renderer not current.");
     impl->line_size = size;
     glLineWidth(size);
+}
+
+void RendererSystem::setMSAA(uint32_t num_samples)
+{
+    impl->msaa_num_samples = num_samples;
+    impl->initFramebuffer(impl->msaa_num_samples, getFramebuffer()->width(), getFramebuffer()->height());
+}
+
+uint32_t RendererSystem::getMSAA() const
+{
+    return impl->msaa_num_samples;
+}
+
+void RendererSystem::toggleMSAA(const bool enable)
+{
+    impl->msaa_enabled = enable;
 }
 
 void RendererSystem::toggleDepthTesting(bool enable)
@@ -1233,21 +1270,13 @@ void RendererSystem::resize(const uint32_t& width, const uint32_t& height)
     ATCG_ASSERT(impl->context->isCurrent(), "Context of Renderer not current.");
 
     setViewport(0, 0, width, height);
-    impl->screen_fbo = atcg::make_ref<Framebuffer>(width, height);
-    impl->screen_fbo->attachColor();
-    TextureSpecification spec_int;
-    spec_int.width  = width;
-    spec_int.height = height;
-    spec_int.format = TextureFormat::RINT;
-    impl->screen_fbo->attachTexture(Texture2D::create(spec_int));
-    impl->screen_fbo->attachDepth();
-    impl->screen_fbo->complete();
+    impl->initFramebuffer(impl->msaa_num_samples, width, height);
 }
 
 void RendererSystem::useScreenBuffer() const
 {
     ATCG_ASSERT(impl->context->isCurrent(), "Context of Renderer not current.");
-    impl->screen_fbo->use();
+    impl->msaa_enabled ? impl->screen_fbo_msaa->use() : impl->screen_fbo->use();
 }
 
 uint32_t RendererSystem::getFrameCounter() const
@@ -1272,10 +1301,12 @@ void RendererSystem::clear() const
     ATCG_ASSERT(impl->context->isCurrent(), "Context of Renderer not current.");
     glClear(impl->clear_flag);
 
-    if(Framebuffer::currentFramebuffer() == impl->screen_fbo->getID())
+    if(Framebuffer::currentFramebuffer() == impl->screen_fbo->getID() ||
+       Framebuffer::currentFramebuffer() == impl->screen_fbo_msaa->getID())
     {
         int value = -1;
         impl->screen_fbo->getColorAttachement(1)->fill(&value);
+        impl->screen_fbo_msaa->getColorAttachement(1)->fill(&value);
     }
 }
 
@@ -1375,7 +1406,7 @@ void RendererSystem::drawCircle(const glm::vec3& position,
 
 void RendererSystem::drawImage(const atcg::ref_ptr<Framebuffer>& img)
 {
-    drawImage(img->getColorAttachement(0));
+    drawImage(std::static_pointer_cast<atcg::Texture2D>(img->getColorAttachement(0)));
 }
 
 void RendererSystem::drawImage(const atcg::ref_ptr<Texture2D>& img)
@@ -1534,6 +1565,11 @@ atcg::ref_ptr<Framebuffer> RendererSystem::getFramebuffer() const
     return impl->screen_fbo;
 }
 
+atcg::ref_ptr<Framebuffer> RendererSystem::getFramebufferMSAA() const
+{
+    return impl->screen_fbo_msaa;
+}
+
 torch::Tensor RendererSystem::getFrame(const torch::DeviceType& device) const
 {
     ATCG_ASSERT(impl->context->isCurrent(), "Context of Renderer not current.");
@@ -1550,7 +1586,7 @@ torch::Tensor RendererSystem::getZBuffer(const torch::DeviceType& device) const
     uint32_t height      = frame->height();
     torch::Tensor buffer = torch::empty({height, width, 1}, atcg::TensorOptions::floatHostOptions());
 
-    useScreenBuffer();
+    impl->screen_fbo->use();
 
     glReadPixels(0, 0, width, height, GL_DEPTH_COMPONENT, GL_FLOAT, buffer.data_ptr());
 
@@ -1561,10 +1597,11 @@ int RendererSystem::getEntityIndex(const glm::vec2& mouse) const
 {
     ATCG_ASSERT(impl->context->isCurrent(), "Context of Renderer not current.");
 
-    useScreenBuffer();
+    impl->screen_fbo->use();
     glReadBuffer(GL_COLOR_ATTACHMENT1);
     int pixelData;
     glReadPixels((int)mouse.x, (int)mouse.y, 1, 1, GL_RED_INTEGER, GL_INT, &pixelData);
+    glReadBuffer(GL_COLOR_ATTACHMENT0);
     return pixelData;
 }
 
@@ -1600,7 +1637,7 @@ void RendererSystem::screenshot(const atcg::ref_ptr<Scene>& scene,
     clear();
     setViewport(0, 0, width, height);
     draw(scene, cam);
-    getFramebuffer()->use();
+    useScreenBuffer();
     setDefaultViewport();
 
     auto data = screenshot_buffer->getColorAttachement(0)->getData(atcg::CPU);
@@ -1626,7 +1663,7 @@ RendererSystem::screenshot(const atcg::ref_ptr<Scene>& scene, const atcg::ref_pt
     clear();
     setViewport(0, 0, width, height);
     draw(scene, cam);
-    getFramebuffer()->use();
+    useScreenBuffer();
     setDefaultViewport();
 
     auto data = screenshot_buffer->getColorAttachement(0)->getData(atcg::CPU);